+++ packages/iov-lisk/src/liskconnection.ts
@@ -13,7 +13,6 @@ import {
   BcpConnection,
   BcpPubkeyQuery,
   BcpQueryEnvelope,
-  BcpQueryTag,
   BcpTicker,
   BcpTransactionResponse,
   BcpTransactionState,
@@ -250,7 +249,7 @@ export class LiskConnection implements BcpConnection {
     }
   }
 
-  public listenTx(_: ReadonlyArray<BcpQueryTag>): Stream<ConfirmedTransaction> {
+  public listenTx(_: BcpTxQuery): Stream<ConfirmedTransaction> {
     throw new Error("Not implemented");
   }