+++ packages/iov-rise/src/riseconnection.ts
@@ -13,7 +13,6 @@ import {
   BcpConnection,
   BcpPubkeyQuery,
   BcpQueryEnvelope,
-  BcpQueryTag,
   BcpTicker,
   BcpTransactionResponse,
   BcpTransactionState,
@@ -268,7 +267,7 @@ export class RiseConnection implements BcpConnection {
     }
   }
 
-  public listenTx(_: ReadonlyArray<BcpQueryTag>): Stream<ConfirmedTransaction> {
+  public listenTx(_: BcpTxQuery): Stream<ConfirmedTransaction> {
     throw new Error("Not implemented");
   }